이미지 줌 구현

+++ app/src/main/java/com/example/fireoperationmap/Place.java
@@ -1,22 +1,22 @@
 package com.example.fireoperationmap;
 
 public class Place {
-    private double x;
-    private double y;
+    private float x;
+    private float y;
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 }
